Remove supporting lnode type for the linked list
Now the linked list is made out of "node" structs.
Was struggling with doing that the first time I implemented the
pairing heap (debugging C is annoying but decided to sit down
and go through the process with gdb). Apparently on merge_pairs
it would end up creating a looping linked list. To solve this
we need to remove the reference to the next node before melding.